+++ src/elapi/cli/__main__.py
@@ -1,4 +1,4 @@
 from .elapi import app
-from .. import APP_NAME
+from .._names import APP_NAME
 
 app(prog_name=APP_NAME)