+++ src/elapi/plugins/commons/__init__.py
@@ -1,6 +1,5 @@
 __all__ = [
     "Typer",
-    "detected_click_feedback",
     "Export",
     "ExportPathValidator",
     "get_structured_data",
@@ -8,12 +7,13 @@ __all__ = [
     "RecursiveInformation",
     "get_location_from_headers",
     "get_whoami",
+    "AsyncInformation"
 ]
 
 
-from .cli_helpers import Typer, detected_click_feedback
+from .cli_helpers import Typer
 from .export import Export, ExportPathValidator
 from .get_data_from_input_or_path import get_structured_data
-from .get_information import Information, RecursiveInformation
+from .get_information import Information, RecursiveInformation, AsyncInformation
 from .get_location_from_headers import get_location_from_headers
 from .get_whoami import get_whoami